+++ types/mempool/mempool.go
@@ -17,6 +17,9 @@ type Mempool interface {
 	// closed by the caller.
 	Select(context.Context, [][]byte) Iterator
 
+	// SelectBy use callback to iterate over the mempool, it's thread-safe to use.
+	SelectBy(context.Context, [][]byte, func(sdk.Tx) bool)
+
 	// CountTx returns the number of transactions currently in the mempool.
 	CountTx() int